Добавили сортировку по режиссёрам

+++ src/main/java/ru/yandex/practicum/filmorate/controller/DirectorController.java
@@ -2,7 +2,14 @@ package ru.yandex.practicum.filmorate.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 import ru.yandex.practicum.filmorate.dto.DirectorDto;
 import ru.yandex.practicum.filmorate.service.DirectorService;
 
@@ -21,7 +28,7 @@ public class DirectorController {
 	}
 
 	@GetMapping("/{id}")
-	public DirectorDto getDirectorById(@PathVariable Integer id) {
+	public DirectorDto getDirectorById(@PathVariable Long id) {
 		return directorService.getDirectorById(id);
 	}
 
@@ -36,7 +43,7 @@ public class DirectorController {
 	}
 
 	@DeleteMapping("/{id}")
-	public void deleteDirector(@PathVariable Integer id) {
+	public void deleteDirector(@PathVariable Long id) {
 		directorService.deleteDirector(id);
 	}
 }
