+++ src/main/java/ru/yandex/practicum/filmorate/dto/DirectorDto.java
@@ -3,6 +3,7 @@ package ru.yandex.practicum.filmorate.dto;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public record DirectorDto(
-	@JsonProperty("id") Integer id,
+	@JsonProperty("id") Long id,
 	@JsonProperty("name") String name
-) {}
+) {
+}
