+++ src/main/java/ru/yandex/practicum/filmorate/storage/director/DirectorStorage.java
@@ -10,13 +10,13 @@ import java.util.Set;
 public interface DirectorStorage {
 	Set<Director> getDirectors();
 
-	Optional<Director> getDirectorById(Integer id);
+	Optional<Director> getDirectorById(Long id);
 
 	Director createDirector(Director director);
 
 	Optional<Director> updateDirector(Director director);
 
-	boolean deleteDirector(Integer id);
+	boolean deleteDirector(Long id);
 
 	void addDirectors(Long filmId, Set<Director> directors);
